+++ packages/@expo/cli/src/export/exportStaticAsync.ts
@@ -33,7 +33,7 @@ type Options = {
   outputDir: string;
   minify: boolean;
   exportServer: boolean;
-  basePath: string;
+  baseUrl: string;
   includeMaps: boolean;
   clear: boolean;
 };
@@ -116,21 +116,22 @@ export async function getFilesToExportFromServerAsync(
 export async function exportFromServerAsync(
   projectRoot: string,
   devServerManager: DevServerManager,
-  { outputDir, basePath, exportServer, minify, includeMaps }: Options
+  { outputDir, baseUrl, exportServer, minify, includeMaps }: Options
 ): Promise<void> {
   const { exp } = getConfig(projectRoot, { skipSDKVersionRequirement: true });
   const appDir = getRouterDirectoryWithManifest(projectRoot, exp);
 
-  const injectFaviconTag = await getVirtualFaviconAssetsAsync(projectRoot, { outputDir, basePath });
+  const injectFaviconTag = await getVirtualFaviconAssetsAsync(projectRoot, { outputDir, baseUrl });
 
   const devServer = devServerManager.getDefaultDevServer();
   assert(devServer instanceof MetroBundlerDevServer);
 
   const [resources, { manifest, serverManifest, renderAsync }] = await Promise.all([
-    devServer.getStaticResourcesAsync({ mode: 'production', minify, includeMaps }),
+    devServer.getStaticResourcesAsync({ mode: 'production', minify, includeMaps, baseUrl }),
     devServer.getStaticRenderFunctionAsync({
       mode: 'production',
       minify,
+      baseUrl,
     }),
   ]);
 
@@ -148,7 +149,7 @@ export async function exportFromServerAsync(
         mode: 'production',
         resources: resources.artifacts,
         template,
-        basePath,
+        baseUrl,
       });
 
       if (injectFaviconTag) {
@@ -170,7 +171,7 @@ export async function exportFromServerAsync(
     await persistMetroAssetsAsync(resources.assets, {
       platform: 'web',
       outputDirectory: outputDir,
-      basePath,
+      baseUrl,
     });
   }
 
@@ -180,6 +181,7 @@ export async function exportFromServerAsync(
       server: devServer,
       appDir,
       manifest: serverManifest,
+      baseUrl,
     });
 
     // Add the api routes to the files to export.
@@ -217,7 +219,7 @@ export function modifyBundlesWithSourceMaps(
   if (filename.endsWith('.js')) {
     // If the bundle ends with source map URLs then update them to point to the correct location.
 
-    // TODO: basePath support
+    // TODO: baseUrl support
     const normalizedFilename = '/' + filename.replace(/^\/+/, '');
     // Ref: https://developer.chrome.com/blog/sourcemaps/#sourceurl-and-displayname-in-action-eval-and-anonymous-functions
     //# sourceMappingURL=//localhost:8085/index.map?platform=web&dev=false&hot=false&lazy=true&minify=true&resolver.environment=client&transform.environment=client&serializer.output=static
@@ -246,17 +248,17 @@ export function getHtmlFiles({
 }): string[] {
   const htmlFiles = new Set<string>();
 
-  function traverseScreens(screens: string | { screens: any; path: string }, basePath = '') {
+  function traverseScreens(screens: string | { screens: any; path: string }, baseUrl = '') {
     for (const value of Object.values(screens)) {
       if (typeof value === 'string') {
-        let filePath = basePath + value;
+        let filePath = baseUrl + value;
         if (value === '') {
           filePath =
-            basePath === ''
+            baseUrl === ''
               ? 'index'
-              : basePath.endsWith('/')
-              ? basePath + 'index'
-              : basePath.slice(0, -1);
+              : baseUrl.endsWith('/')
+              ? baseUrl + 'index'
+              : baseUrl.slice(0, -1);
         }
         if (includeGroupVariations) {
           // TODO: Dedupe requests for alias routes.
@@ -265,7 +267,7 @@ export function getHtmlFiles({
           htmlFiles.add(filePath);
         }
       } else if (typeof value === 'object' && value?.screens) {
-        const newPath = basePath + value.path + '/';
+        const newPath = baseUrl + value.path + '/';
         traverseScreens(value.screens, newPath);
       }
     }
@@ -346,12 +348,14 @@ async function exportApiRoutesAsync({
   outputDir,
   server,
   appDir,
+  baseUrl,
   ...props
 }: {
   outputDir: string;
   server: MetroBundlerDevServer;
   appDir: string;
   manifest: ExpoRouterServerManifestV1;
+  baseUrl: string;
 }): Promise<Map<string, string>> {
   const functionsDir = '_expo/functions';
   const funcDir = path.join(outputDir, functionsDir);
@@ -362,6 +366,7 @@ async function exportApiRoutesAsync({
     appDir,
     outputDir: functionsDir,
     prerenderManifest: props.manifest,
+    baseUrl,
   });
 
   Log.log(chalk.bold`Exporting ${files.size} API Routes.`);