+++ packages/@expo/cli/src/start/server/middleware/ManifestMiddleware.ts
@@ -11,7 +11,11 @@ import path from 'path';
 import { resolve } from 'url';
 
 import { ExpoMiddleware } from './ExpoMiddleware';
-import { shouldEnableAsyncImports, createBundleUrlPath } from './metroOptions';
+import {
+  shouldEnableAsyncImports,
+  createBundleUrlPath,
+  getBaseUrlFromExpoConfig,
+} from './metroOptions';
 import { resolveGoogleServicesFile, resolveManifestAssets } from './resolveAssets';
 import { parsePlatformHeader, RuntimePlatform } from './resolvePlatform';
 import { ServerHeaders, ServerNext, ServerRequest, ServerResponse } from './server.types';
@@ -160,6 +164,7 @@ export abstract class ManifestMiddleware<
       mainModuleName,
       hostname,
       engine: isHermesEnabled ? 'hermes' : undefined,
+      baseUrl: getBaseUrlFromExpoConfig(projectConfig.exp),
     });
 
     // Resolve all assets and set them on the manifest as URLs
@@ -211,11 +216,13 @@ export abstract class ManifestMiddleware<
     mainModuleName,
     hostname,
     engine,
+    baseUrl,
   }: {
     platform: string;
     hostname?: string | null;
     mainModuleName: string;
     engine?: 'hermes';
+    baseUrl?: string;
   }): string {
     const path = createBundleUrlPath({
       mode: this.options.mode ?? 'development',
@@ -224,6 +231,7 @@ export abstract class ManifestMiddleware<
       mainModuleName,
       lazy: shouldEnableAsyncImports(this.projectRoot),
       engine,
+      baseUrl,
     });
 
     return (
@@ -307,6 +315,7 @@ export abstract class ManifestMiddleware<
       mode: this.options.mode ?? 'development',
       // Hermes doesn't support more modern JS features than most, if not all, modern browser.
       engine: 'hermes',
+      baseUrl: getBaseUrlFromExpoConfig(this.initialProjectConfig.exp),
     });
   }